+++ src/js/components/Meta/index.ts
@@ -1,0 +1,5 @@
+import { withRouter } from 'react-router';
+
+import Meta from './Meta';
+
+export default withRouter(Meta);
