+++ src/js/components/ShabadContent/index.ts
@@ -1,1 +1,8 @@
-export { default } from './ShabadContent';
+import { connect } from 'react-redux';
+import { IStore } from '@/features/types';
+
+import ShabadContent from './ShabadContent';
+
+const stateToProps = (state: IStore) => state;
+
+export default connect(stateToProps)(ShabadContent);
