--- src/js/components/index.d.ts
@@ -1,1 +1,0 @@
-type ShabadContentTypes = 'shabad' | 'ang' | 'hukamnama';
